move functionality into appropriate classes
